+++ app/src/main/java/com/criteo/testapp/StandaloneActivity.java
@@ -33,7 +33,6 @@ import com.criteo.publisher.integration.Integration;
 import com.criteo.publisher.model.InterstitialAdUnit;
 import com.criteo.testapp.integration.MockedIntegrationRegistry;
 import com.criteo.testapp.listener.TestAppBannerAdListener;
-import com.criteo.testapp.listener.TestAppInterstitialAdDisplayListener;
 import com.criteo.testapp.listener.TestAppInterstitialAdListener;
 import com.criteo.testapp.listener.TestAppNativeAdListener;
 
@@ -117,11 +116,9 @@ public class StandaloneActivity extends AppCompatActivity {
   private void loadInterstitial(InterstitialAdUnit adUnit, Button btnShow) {
     String prefix = "Standalone " + adUnit.getAdUnitId();
 
-    CriteoInterstitial criteoInterstitial = new CriteoInterstitial(context, adUnit);
+    CriteoInterstitial criteoInterstitial = new CriteoInterstitial(adUnit);
     criteoInterstitial.setCriteoInterstitialAdListener(
         new TestAppInterstitialAdListener(TAG, prefix, btnShow));
-    criteoInterstitial.setCriteoInterstitialAdDisplayListener(
-        new TestAppInterstitialAdDisplayListener(TAG, prefix));
 
     btnShow.setOnClickListener(v -> showInterstitial(criteoInterstitial));