+++ publisher-sdk/src/androidTest/java/com/criteo/publisher/integration/InHouseFunctionalTest.java
@@ -132,7 +132,7 @@ public class InHouseFunctionalTest {
     interstitial.loadAd(bidResponse.getBidToken());
     waitForIdleState();
 
-    verify(listener).onAdReceived();
+    verify(listener).onAdReceived(interstitial);
 
     verify(api, atLeastOnce()).loadCdb(
         argThat(request -> request.getProfileId() == Integration.IN_HOUSE.getProfileId()),
@@ -167,11 +167,13 @@ public class InHouseFunctionalTest {
 
     BidResponse bidResponse = criteo.getBidResponse(validInterstitialAdUnit);
     interstitial.loadAd(bidResponse.getBidToken());
+    waitForIdleState();
+
     interstitial.loadAd(bidResponse.getBidToken());
     waitForIdleState();
 
     InOrder inOrder = inOrder(listener);
-    inOrder.verify(listener).onAdReceived();
+    inOrder.verify(listener).onAdReceived(interstitial);
     inOrder.verify(listener).onAdFailedToReceive(ERROR_CODE_NO_FILL);
     inOrder.verifyNoMoreInteractions();
   }
@@ -186,7 +188,7 @@ public class InHouseFunctionalTest {
   @NonNull
   private CriteoInterstitial createInterstitial(InterstitialAdUnit interstitialAdUnit) {
     return callOnMainThreadAndWait(() ->
-        new CriteoInterstitial(context, interstitialAdUnit)
+        new CriteoInterstitial(interstitialAdUnit)
     );
   }