+++ publisher-sdk/src/main/java/com/criteo/publisher/model/WebViewData.java
@@ -17,9 +17,9 @@
 package com.criteo.publisher.model;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import com.criteo.publisher.CriteoInterstitialAdDisplayListener;
 import com.criteo.publisher.DependencyProvider;
+import com.criteo.publisher.network.PubSdkApi;
+import com.criteo.publisher.tasks.InterstitialListenerNotifier;
 import com.criteo.publisher.tasks.WebViewDataTask;
 import com.criteo.publisher.util.WebViewLoadStatus;
 import java.util.concurrent.Executor;
@@ -35,10 +35,14 @@ public class WebViewData {
   @NonNull
   private final Config config;
 
-  public WebViewData(@NonNull Config config) {
+  @NonNull
+  private final PubSdkApi api;
+
+  public WebViewData(@NonNull Config config, @NonNull PubSdkApi api) {
     this.content = "";
     this.webViewLoadStatus = WebViewLoadStatus.NONE;
     this.config = config;
+    this.api = api;
   }
 
   public boolean isLoaded() {
@@ -79,9 +83,18 @@ public class WebViewData {
   public void fillWebViewHtmlContent(
       @NonNull String displayUrl,
       @NonNull DeviceInfo deviceInfo,
-      @Nullable CriteoInterstitialAdDisplayListener criteoInterstitialAdDisplayListener) {
+      @NonNull InterstitialListenerNotifier listenerNotifier
+  ) {
     Executor threadPoolExecutor = DependencyProvider.getInstance().provideThreadPoolExecutor();
-    new WebViewDataTask(this, deviceInfo, criteoInterstitialAdDisplayListener)
-        .executeOnExecutor(threadPoolExecutor, displayUrl);
+
+    Runnable task = new WebViewDataTask(
+        displayUrl,
+        this,
+        deviceInfo,
+        listenerNotifier,
+        api
+    );
+
+    threadPoolExecutor.execute(task);
   }
 }