+++ publisher-sdk/src/test/java/com/criteo/publisher/advancednative/RendererHelperTest.kt
@@ -19,7 +19,13 @@ package com.criteo.publisher.advancednative
 import android.graphics.drawable.Drawable
 import android.widget.ImageView
 import com.criteo.publisher.concurrent.DirectMockRunOnUiThreadExecutor
-import com.nhaarman.mockitokotlin2.*
+import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.doAnswer
+import com.nhaarman.mockitokotlin2.doReturn
+import com.nhaarman.mockitokotlin2.doThrow
+import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.verify
+import com.nhaarman.mockitokotlin2.whenever
 import org.assertj.core.api.Assertions.assertThatCode
 import org.junit.Before
 import org.junit.Test
@@ -80,7 +86,7 @@ class RendererHelperTest {
     helper.setMediaInView(media, mediaView)
 
     verify(imageLoader).loadImageInto(url, imageView, placeholder)
-    uiExecutor.expectIsRunningInExecutor()
+    uiExecutor.verifyExpectations()
   }
 
   @Test
@@ -91,5 +97,4 @@ class RendererHelperTest {
       helper.setMediaInView(mock(), mock())
     }.doesNotThrowAnyException()
   }
-
 }